Move recurse_status_for_sample to utils/charon.py
